Added rudimentary log entry saving functionality
Need to debug still - issue with the first-selected item not showing as selected (though this does appear to be saving in the app DB). Also need to refine database reset so that the displayed log entries always correspond to the selected log filter

+++ app/src/main/java/com/scriptient/rxplorer/async/DatabaseResetAsyncTask.java
@@ -27,7 +27,11 @@ public class DatabaseResetAsyncTask extends AsyncTask<Void, Void, Void> {
 
         for ( AppEmbeddedLogEntry entry : allEntries ) {
 
-            AppEmbeddedLogEntryRepo.deleteLogEntry( viewWeakReference.get().getContext(), entry );
+            if ( !entry.getSaved() ) {
+
+                AppEmbeddedLogEntryRepo.deleteLogEntry( viewWeakReference.get().getContext(), entry );
+
+            }
 
         }
